ai-model/requirements.txt
pandas==2.1.4
sentence-transformers==2.2.2
chromadb==0.4.22
google-generativeai==0.3.2
openai==1.6.1
numpy==1.24.3
python-dotenv==1.0.0
tqdm==4.66.1
scikit-learn==1.3.2
